wrap app with StaticRouter from server side and BrowserRouter from client side
iWe pass the current url (Express’ req.url) to the location prop and an empty object to the context prop. The context object is useful to store information about a specific route render, and that information is then made available to the component in the form of a staticContext prop.